+++ src/HeavyGear.Core/GameLogic/Vehicle.cs
@@ -763,13 +763,13 @@ namespace HeavyGear.GameLogic
 
                 if (target.ReinforcedFrontArmor > 0)
                 {
-                    if (CheckDefenseArc((int)target.Rotation, MapPosition, target.MapPosition) == 0)
+                    if (CheckDefenseArc(target.Rotation, MapPosition, target.MapPosition) == 0)
                         effectiveArmor += target.ReinforcedFrontArmor;
                 }
 
                 if (target.WeakRearFacing)
                 {
-                    if (CheckDefenseArc((int)target.Rotation, MapPosition, target.MapPosition) > 0)
+                    if (CheckDefenseArc(target.Rotation, MapPosition, target.MapPosition) > 0)
                         effectiveArmor /= 2;
                 }
 
@@ -906,7 +906,7 @@ namespace HeavyGear.GameLogic
             #region Defender Modifiers
 
             //defense arc modifier
-            int defenseArcModifier = CheckDefenseArc((int)target.Rotation, MapPosition, target.MapPosition);
+            int defenseArcModifier = CheckDefenseArc(target.Rotation, MapPosition, target.MapPosition);
 
             #region Apply Modifiers
